+++ src/main/java/com/mumuk/domain/healthManagement/dto/response/UserInfoResponse.java
@@ -10,8 +10,8 @@ public class UserInfoResponse {
     @AllArgsConstructor
     public static class UserInfoRes {
         private Gender gender;
-        private Long height;
-        private Long weight;
+        private Double height;
+        private Double weight;
     }
 
 }
